requirements.txt
# Web
Flask>=3.0.0
Werkzeug>=3.0.0

# WebSocket
flask-sock>=0.7.0
gevent>=23.9.0

# Production server
gunicorn>=21.0.0
gevent-websocket>=0.10.0

# HTTP client
requests>=2.31.0
httpx>=0.25.0

# Text extraction (no external API)
PyMuPDF>=1.23.0
python-docx>=1.1.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Audio
pydub>=0.25.1

# OpenAPI clients (custom URL + API_KEY)
openai>=1.0.0

# DB
# SQLite via sqlite3 stdlib

# Utils
python-dotenv>=1.0.0
mutagen>=1.47.0

# Tests
pytest>=7.4.0
